Added some forgotten logic.
I noticed that we never used conditional logic to check whether
we were supposed to display site visits (session) data by region
or country. Country was just hard-coded in. I have made it check
for that and tested with region and country options. Lastly, I
realized that we forgot to hide users online when show_users_online
is false in the UsersOpts model. I have resolved and tested that as
well.